requirements.txt
textblob
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
azure-common==1.1.28
azure-core==1.36.0
azure-identity==1.25.1
azure-search-documents==11.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
exceptiongroup==1.3.1
fastapi==0.124.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
msal==1.34.0
msal-extensions==1.3.1
numpy==2.2.6
pandas==2.3.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.38.0
